refactor: add refactor of originally generated video call screen
